+++ objects/src/accounts/account_id.rs
@@ -20,13 +20,13 @@ use crate::{
 // ACCOUNT TYPE
 // ================================================================================================
 
-const FUNGIBLE_FAUCET: u64 = 0b10;
-const NON_FUNGIBLE_FAUCET: u64 = 0b11;
-const REGULAR_ACCOUNT_IMMUTABLE_CODE: u64 = 0b00;
-const REGULAR_ACCOUNT_UPDATABLE_CODE: u64 = 0b01;
+const FUNGIBLE_FAUCET: u8 = 0b10;
+const NON_FUNGIBLE_FAUCET: u8 = 0b11;
+const REGULAR_ACCOUNT_IMMUTABLE_CODE: u8 = 0b00;
+const REGULAR_ACCOUNT_UPDATABLE_CODE: u8 = 0b01;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-#[repr(u64)]
+#[repr(u8)]
 pub enum AccountType {
     FungibleFaucet = FUNGIBLE_FAUCET,
     NonFungibleFaucet = NON_FUNGIBLE_FAUCET,
@@ -60,14 +60,74 @@ impl rand::distributions::Distribution<AccountType> for rand::distributions::Sta
     }
 }
 
+// SERIALIZATION
+// ================================================================================================
+
+impl Serializable for AccountType {
+    fn write_into<W: vm_core::utils::ByteWriter>(&self, target: &mut W) {
+        target.write_u8(*self as u8);
+    }
+}
+
+impl Deserializable for AccountType {
+    fn read_from<R: ByteReader>(source: &mut R) -> Result<Self, DeserializationError> {
+        let num: u8 = source.read()?;
+        match num {
+            FUNGIBLE_FAUCET => Ok(AccountType::FungibleFaucet),
+            NON_FUNGIBLE_FAUCET => Ok(AccountType::NonFungibleFaucet),
+            REGULAR_ACCOUNT_IMMUTABLE_CODE => Ok(AccountType::RegularAccountImmutableCode),
+            REGULAR_ACCOUNT_UPDATABLE_CODE => Ok(AccountType::RegularAccountUpdatableCode),
+            _ => Err(DeserializationError::InvalidValue(format!("invalid account type: {num}"))),
+        }
+    }
+}
+
+#[cfg(feature = "std")]
+impl serde::Serialize for AccountType {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        let s = match self {
+            AccountType::FungibleFaucet => "FungibleFaucet",
+            AccountType::NonFungibleFaucet => "NonFungibleFaucet",
+            AccountType::RegularAccountImmutableCode => "RegularAccountImmutableCode",
+            AccountType::RegularAccountUpdatableCode => "RegularAccountUpdatableCode",
+        };
+        serializer.serialize_str(s)
+    }
+}
+
+#[cfg(feature = "std")]
+impl<'de> serde::Deserialize<'de> for AccountType {
+    fn deserialize<D>(deserializer: D) -> Result<AccountType, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        use serde::de::Error;
+        let s: String = serde::Deserialize::deserialize(deserializer)?;
+
+        match s.as_str() {
+            "FungibleFaucet" => Ok(AccountType::FungibleFaucet),
+            "NonFungibleFaucet" => Ok(AccountType::NonFungibleFaucet),
+            "RegularAccountImmutableCode" => Ok(AccountType::RegularAccountImmutableCode),
+            "RegularAccountUpdatableCode" => Ok(AccountType::RegularAccountUpdatableCode),
+            other => Err(D::Error::invalid_value(
+                serde::de::Unexpected::Str(other),
+                &"a valid account type (\"FungibleFaucet\", \"NonFungibleFaucet\", \"RegularAccountImmutableCode\", or \"RegularAccountUpdatableCode\")",
+            )),
+        }
+    }
+}
+
 // ACCOUNT STORAGE MODE
 // ================================================================================================
 
-const PUBLIC: u64 = 0b00;
-const PRIVATE: u64 = 0b10;
+const PUBLIC: u8 = 0b00;
+const PRIVATE: u8 = 0b10;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[repr(u64)]
+#[repr(u8)]
 pub enum AccountStorageMode {
     Public = PUBLIC,
     Private = PRIVATE,
@@ -264,7 +324,7 @@ impl AccountId {
 
     /// The lower two bits of the second least significant nibble determine the account type.
     pub(crate) const TYPE_SHIFT: u64 = 4;
-    pub(crate) const TYPE_MASK: u64 = 0b11 << Self::TYPE_SHIFT;
+    pub(crate) const TYPE_MASK: u8 = 0b11 << Self::TYPE_SHIFT;
 
     /// The least significant nibble determines the account version.
     const VERSION_MASK: u64 = 0b1111;
@@ -275,7 +335,7 @@ impl AccountId {
     /// The higher two bits of the second least significant nibble determine the account storage
     /// mode.
     pub(crate) const STORAGE_MODE_SHIFT: u64 = 6;
-    pub(crate) const STORAGE_MODE_MASK: u64 = 0b11 << Self::STORAGE_MODE_SHIFT;
+    pub(crate) const STORAGE_MODE_MASK: u8 = 0b11 << Self::STORAGE_MODE_SHIFT;
 
     pub(crate) const IS_FAUCET_MASK: u64 = 0b10 << Self::TYPE_SHIFT;
 
@@ -700,8 +760,10 @@ fn validate_second_felt(second_felt: Felt) -> Result<(), AccountError> {
 }
 
 pub(crate) fn extract_storage_mode(first_felt: u64) -> Result<AccountStorageMode, AccountError> {
-    let bits = (first_felt & AccountId::STORAGE_MODE_MASK) >> AccountId::STORAGE_MODE_SHIFT;
-    match bits {
+    let bits =
+        (first_felt & (AccountId::STORAGE_MODE_MASK as u64)) >> AccountId::STORAGE_MODE_SHIFT;
+    // SAFETY: `STORAGE_MODE_MASK` is u8 so casting bits is lossless
+    match bits as u8 {
         PUBLIC => Ok(AccountStorageMode::Public),
         PRIVATE => Ok(AccountStorageMode::Private),
         _ => Err(AccountError::InvalidAccountStorageMode(format!("0b{bits:b}"))),
@@ -720,8 +782,9 @@ pub(crate) fn extract_version(first_felt: u64) -> Result<AccountIdVersion, Accou
 }
 
 pub(crate) const fn extract_type(first_felt: u64) -> AccountType {
-    let bits = (first_felt & AccountId::TYPE_MASK) >> AccountId::TYPE_SHIFT;
-    match bits {
+    let bits = (first_felt & (AccountId::TYPE_MASK as u64)) >> AccountId::TYPE_SHIFT;
+    // SAFETY: `TYPE_MASK` is u8 so casting bits is lossless
+    match bits as u8 {
         REGULAR_ACCOUNT_UPDATABLE_CODE => AccountType::RegularAccountUpdatableCode,
         REGULAR_ACCOUNT_IMMUTABLE_CODE => AccountType::RegularAccountImmutableCode,
         FUNGIBLE_FAUCET => AccountType::FungibleFaucet,
@@ -938,7 +1001,7 @@ mod tests {
     /// normal.
     #[test]
     fn test_account_id_faucet_bit() {
-        const ACCOUNT_IS_FAUCET_MASK: u64 = 0b10;
+        const ACCOUNT_IS_FAUCET_MASK: u8 = 0b10;
 
         // faucets have a bit set
         assert_ne!((FUNGIBLE_FAUCET) & ACCOUNT_IS_FAUCET_MASK, 0);