+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Added metrics to the `miden-tx-prover` proxy (#1017).
 - Implemented `to_hex` for `AccountIdPrefix` and `epoch_block_num` for `BlockHeader` (#1039).
 - Introduce `AccountIdBuilder` to simplify `AccountId` generation in tests (#1045).
+- Introduced `AccountComponentTemplate` with TOML serialization and templating (#1015, #1027).
 
 ## 0.6.2 (2024-11-20)